requirements.txt
numpy~=1.26.4
pillow~=10.3.0
tqdm~=4.66.4
torchvision~=0.15.2
packaging~=24.0
peft~=0.10.0
diffusers~=0.25.1
transformers~=4.35.2
wandb~=0.17.0
lpips~=0.1.4
accelerate~=0.30.0